Prepare the split of arpifs4py into pyfa(&lfi) and transforms (later in ectrans)
