+++ epygram/geometries/SpectralGeometry.py
@@ -173,10 +173,10 @@ class SpectralGeometry(RecursiveObject):
         call.
         """
         if not hasattr(cls, '_transforms_lib'):
-            from arpifs4py import wtransforms, init_env
+            from arpifs4py import wtransforms
             cls._transforms_lib = wtransforms
             if transforms_lib_init_env_kwargs.pop('trigger', False):
-                init_env(**transforms_lib_init_env_kwargs)
+                wtransforms.init_env(**transforms_lib_init_env_kwargs)
         return cls._transforms_lib
 
     def _prevent_swapping(self):
